video-creator/requirements.txt
pika
pydantic
python-dotenv
requests
moviepy==1.0.3
numpy
Pillow
imageio-ffmpeg
